requirements.txt
pillow==9.2.0
numpy==1.23.4
matplotlib==3.6.1
tensorflow==2.*
scikit-learn==1.1.3
tensorflow_hub==0.12.0
six==1.16.0
keras==2.*